29 jan

+++ Algorithms/Monthlypayment.php
@@ -1,5 +1,6 @@
 <?php
 /**
+* File:Monthlypayment.php
 * Purpose: program to calculate the monthly payment by taking principal as(p) ,year as(y) and rate as(r)
 * @author :Nishithkumar
 * @version :1.0
@@ -13,6 +14,6 @@ $p=Alutility::getInt();
 echo " enter how many years:";
 $y=Alutility::getInt();
 echo "enter the rate value:";
-$r=Alutility::getInt();
+$r=Alutility::getFloat();
 Alutility::monthlypayment($p,$y,$r); // caling the function declared in utility
 ?>
